plugins/magellon_result_processor/requirements.txt
fastapi
uvicorn
pydantic 
rich
pyyaml

python-json-logger


prometheus_client
python-dotenv
# elastic-apm

python-consul
pika
# aio_pika

sqlalchemy
sqlalchemy_utils
PyMySQL==1.1.0
#sqlacodegen
mrcfile
#Pillow==7.0.0 #for ctf evaluation the syntax supports only this version
numpy
scipy
Pillow
matplotlib

prometheus-fastapi-instrumentator

websockets
debugpy